[FIZZBUZZ] Using a Moderator class
Notice the use of (1..100).each, this is another method of iterating over a set
of objects that you should prefer over for-loops

This is definitely overkill for this problem, but should serve as an example of
what you can do with classes

Also note that this is **not** the best use-case for a class because we really
don't encapsulate internal state in the Moderator. A module might be more
appropriate or a generic function.